resolves #1